+++ test-workspace/CSharpLib/Class1.cs
@@ -7,7 +7,7 @@ public class GreeterService
 		return $"Hello, {name} from C#";
 	}
 
-	public int Add(int left, int right)
+	public int SAdd(int left, int right)
 	{
 		return left + right;
 	}